Fix compat-table failures: Array.at, AnnexB if-decl, Proxy observable gets
- Fix Array.prototype.at() returning wrong value for negative out-of-bounds
  indices by changing ulong to long (C# casts negative double to ulong as 0)
- Fix AnnexB if-decl-else-decl with same function name in both branches by
  not adding to effectiveLexicalNames for IfStatement siblings
- Use RegExpCreate instead of RegExp.Construct in String.prototype.match/search
  to avoid extra IsRegExp call (observable Symbol.match access through Proxy)
- Remove observable sticky property read in RegExp.prototype[Symbol.replace]
  fast path by deriving from already-read flags string
- Fix String.prototype.split to use direct JsRegExp check instead of IsRegExp
  to avoid observable Symbol.match access

+++ Jint/Native/Array/ArrayPrototype.cs
@@ -890,22 +890,22 @@ public sealed class ArrayPrototype : ArrayInstance
         var len = target.GetLength();
         var relativeIndex = TypeConverter.ToInteger(arguments.At(0));
 
-        ulong actualIndex;
+        long actualIndex;
         if (relativeIndex < 0)
         {
-            actualIndex = (ulong) (len + relativeIndex);
+            actualIndex = (long) (len + relativeIndex);
         }
         else
         {
-            actualIndex = (ulong) relativeIndex;
+            actualIndex = (long) relativeIndex;
         }
 
-        if (actualIndex < 0 || actualIndex >= len)
+        if (actualIndex < 0 || (ulong) actualIndex >= len)
         {
             return Undefined;
         }
 
-        return target.Get(actualIndex);
+        return target.Get((ulong) actualIndex);
     }
 
     /// <summary>
